+++ src/gui.h
@@ -30,10 +30,10 @@ struct UIState
 
 extern struct UIState uistate;
 
-void DrawRect(uint8_t x, uint8_t y, uint8_t w, uint8_t h, COLOR color);
+void DrawRect(uint16_t x, uint16_t y, uint16_t w, uint16_t h, COLOR color);
 void DrawString(uint16_t x, uint16_t y, uint8_t *pStr, COLOR textColor, COLOR backColor);
 
-uint8_t RegionHit(uint8_t x, uint8_t y, uint8_t w, uint8_t h);
+uint8_t RegionHit(uint16_t x, uint16_t y, uint16_t w, uint16_t h);
 void Imgui_Prepare();
 void Imgui_Finish();
